requirements.txt
anyio==4.9.0
APScheduler==3.11.0
certifi==2025.1.31
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
python-telegram-bot==22.0
sniffio==1.3.1
tzdata==2025.1
tzlocal==5.3.1
